requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
assemblyai==0.35.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Automat==24.8.1
Babel==2.15.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
build==1.2.2.post1
cachetools==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.8
constantly==23.10.4
contourpy==1.2.1
cryptography==44.0.0
cssselect==1.2.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
durationpy==0.9
elevenlabs==1.13.2
executing==2.0.1
faiss-cpu==1.9.0.post1
fastapi==0.115.6
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.4.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.26.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.4
importlib-metadata==6.11.0
importlib_resources==6.4.5
imutils==0.5.4
incremental==24.7.2
InstructorEmbedding==1.0.1
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
keras==3.6.0
kiwisolver==1.4.5
kubernetes==31.0.0
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-groq==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.4
libclang==18.1.1
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mltu==1.2.5
mmh3==5.0.1
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.56.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.12
overrides==7.7.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
Pillow==10.0.1
platformdirs==4.2.2
posthog==3.7.4
proglog==0.1.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
Protego==0.3.1
proto-plus==1.25.0
protobuf==4.25.8
psutil==6.0.0
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.7.0
pydantic_core==2.27.1
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.18.0
PyMuPDF==1.26.0
pyOpenSSL==24.3.0
pyparsing==3.1.2
PyPDF2==3.0.1
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-pptx==0.6.21
pyttsx3==2.98
pytube==15.0.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qqdm==0.0.7
qtconsole==5.5.2
QtPy==2.4.1
queuelib==1.7.0
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.13.1
Scrapy==2.12.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
service-identity==24.2.0
setuptools==70.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.11.0
speedtest-cli==2.1.3
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.29.0
sympy==1.13.1
tenacity==8.5.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
toolz==0.12.1
torch==2.5.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.47.1
Twisted==24.11.0
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.34.0
validators==0.35.0
w3lib==2.2.1
watchdog==4.0.1
watchfiles==1.0.3
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.11
wikipedia==1.4.0
Wikipedia-API==0.6.0
wordcloud==1.9.3
wrapt==1.16.0
xgboost==2.1.3
XlsxWriter==3.2.3
yarl==1.9.4
yt-dlp==2024.12.13
zipp==3.21.0
zope.interface==7.2
